Former-commit-id: d3e7a8dfad74b255d77b24e425c3b47e3184d0b2
Former-commit-id: 19977d2bb667ef39591292157a852c58547b8c19

+++ p2p/p2pserver.go
@@ -424,7 +424,7 @@ func (s *P2pServer) ServerStreamRead(stream pb.P2Pgservice_ServerStreamReadServe
 			}
 
 			log.Info("ServerStreamRead", " Recv block==+=====+=>Height", block.GetBlock().GetHeight(),
-				"block size(KB)", len(pb.Encode(block))/1024, "block hash", blockhash)
+				"block size(KB)", float32(len(pb.Encode(block)))/1024, "block hash", blockhash)
 			if block.GetBlock() != nil {
 				//msg := s.node.nodeInfo.client.NewMessage("blockchain", pb.EventBroadcastAddBlock, block.GetBlock())
 				msg := s.node.nodeInfo.client.NewMessage("blockchain", pb.EventBroadcastAddBlock, &pb.BlockPid{peername, block.GetBlock()})
